requirements.txt
Django==2.1
django-mptt==0.9.1
djangorestframework==3.8.2
flake8==3.5.0
ipython==6.5.0
psycopg2-binary==2.7.5
Pillow==5.2.0
transliterate==1.10.1